+++ CHANGELOG.md
@@ -15,8 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Column title has been renamed form "Display Number" to "ID"
 - Move `default_due` config to core
 - Move `default_priority` config to core
+- Move `short_id` to core
 
 ## [v0.2.2] - 2025-07-28